+++ src/main/java/ma/vi/esql/builder/DeleteBuilder.java
@@ -144,6 +144,6 @@ public class DeleteBuilder implements Builder<Delete> {
   private final List<Attribute> returnMetadata = new ArrayList<>();
   private final List<Column> returnColumns = new ArrayList<>();
 
-  private final Context context;
-  private final Parser parser;
+  public final Context context;
+  public final Parser parser;
 }