+++ src/main/java/ma/vi/esql/builder/UpdateBuilder.java
@@ -156,6 +156,6 @@ public class UpdateBuilder implements Builder<Update> {
   private final List<Attribute> returnMetadata = new ArrayList<>();
   private final List<Column> returnColumns = new ArrayList<>();
 
-  private final Context context;
-  private final Parser parser;
+  public final Context context;
+  public final Parser parser;
 }